+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 - **Added**: Support let tag for inline declarations
+- **Added**: Support match statements
 
 ## [0.3.0](https://codeberg.org/lsunsi/hiper/src/tag/0.3.0)
 - **Added**: Raw struct (to bypass escaping)